backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.8.0
pydantic-settings>=2.3.0
gitpython>=3.1.43
PyDriller>=2.6
networkx>=3.3
pytest>=8.2.0
httpx>=0.27.0
